+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `ui/Touchable` to end gestures when focus is lost
+
 ## [2.0.0-beta.4] - 2018-05-21
 
 ### Fixed